+++ state/actions.js
@@ -96,6 +96,14 @@ export function updateSubstitutionAdvanceWidth(value) {
   }
 }
 
+export function updateSubstitutionDiacritic(value) {
+  return {
+    type: UPDATE_SUBSTITUTION,
+    field: 'glyph/diacritic',
+    value
+  }
+}
+
 export function removeSubstitution({ active }) {
   return dispatch => {
     dispatch({
@@ -157,7 +165,7 @@ export function swapSubstitution(a, b) {
 }
 
 export function save({ fonts: { substitutions, fontname }}) {
-  const data = 'data:text/json;charset=utf-8,' + encodeURIComponent(JSON.stringify({ version: '0.1.0', substitutions, fontname }))
+  const data = 'data:text/json;charset=utf-8,' + encodeURIComponent(JSON.stringify({ version: '0.1.1', substitutions, fontname }))
   const container = document.createElement('a')
   container.setAttribute('href', data)
   container.setAttribute('download', slugify(fontname || 'conscripter-custom-font') + '.json')